+++ examples/grid_factory.rs
@@ -50,7 +50,7 @@ impl FactoryComponent for Counter {
     type ParentWidget = gtk::Grid;
     type Index = DynamicIndex;
 
-    fn output_to_parent_input(output: Self::Output) -> Option<AppMsg> {
+    fn forward_to_parent(output: Self::Output) -> Option<AppMsg> {
         Some(match output {
             CounterOutput::SendFront(index) => AppMsg::SendFront(index),
             CounterOutput::MoveUp(index) => AppMsg::MoveUp(index),