+++ relm4/src/factory/async/component_storage.rs
@@ -70,7 +70,7 @@ where
                 index,
                 returned_widget,
                 parent_sender,
-                C::output_to_parent_input,
+                C::forward_to_parent,
             )))
         } else {
             None