+++ relm4/src/factory/sync/traits.rs
@@ -29,7 +29,7 @@ pub trait FactoryComponent:
     /// The parameter used to initialize the factory component.
     type Init;
 
-    /// The widget that was constructed by the factory component.
+    /// The top-level widget of the factory component.
     type Root: AsRef<<Self::ParentWidget as FactoryView>::Children> + Debug + Clone;
 
     /// The type that's used for storing widgets created for this factory component.
@@ -63,7 +63,7 @@ pub trait FactoryComponent:
     /// forward messages.
     ///
     /// If [`None`] is returned, nothing is forwarded.
-    fn output_to_parent_input(_output: Self::Output) -> Option<Self::ParentInput> {
+    fn forward_to_parent(_output: Self::Output) -> Option<Self::ParentInput> {
         None
     }