+++ CHANGES.md
@@ -9,11 +9,13 @@
 + core: Documentation and better support for data bindings
 + core: Add `set_tooltip` method to `RelmWidgetExt`
 + core: Add `main_adw_application` method to retrieve the `adw::Application` when the libadwaita feature is enabled
++ macros: Add `skip_macro` option for watch and track attributes to skip their initialization 
 + examples: Example for using relm4-icons 
 
 ### Changed
 
 + core: Added `Index` type to the `FactoryComponent` trait
++ core: Rename factory component traits `output_to_parent_input` method to `forward_to_parent`
 + core: Improved `RelmActionGroup` API
 + all: Increase MSRV to 1.65 to match the dependencies